♻️ | refac(home-page): renomeia homePage e refatora pra deixar dinâmico independente do tipo de usuário logado
